requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.94.1
h11==0.14.0
idna==3.4
numpy==1.24.2
pandas==1.3.3
pydantic==1.10.6
pymongo==3.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
uvicorn==0.21.1
